moving delivery type folder
